All the types pass
