+++ src/utilities/index.ts
@@ -6,6 +6,7 @@ export * from './components'
 export * from './copy'
 export * from './dates'
 export * from './delete'
+export * from './functions'
 export * from './httpStatus'
 export * from './inject'
 export * from './json'
